requirements.txt
flask==2.3.3
pdf2image==1.16.3
google-genai>=1.0.0
python-dotenv==1.0.0
Pillow>=10.0.0
Werkzeug==2.3.7 